% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Landmark Data}
\format{
A list of specimen landmark curves.
}
\usage{
example_data
}
\description{
A small subset of raw landmark data for testing and demonstration
(data after using \code{extract_landmarks} function)
}
\examples{
data("example_data", package = "AmmoniTools")
str(example_data)
}
\keyword{datasets}
