% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{plot_cac_shapes3d}
\alias{plot_cac_shapes3d}
\title{Plots 3D shapes along a CAC axis}
\usage{
plot_cac_shapes3d(
  gpa,
  pca,
  cac_res,
  min_col = "blue",
  max_col = "red",
  mid_col = "black",
  size_min = 1.5,
  size_mid = 1.5,
  size_max = 1.5,
  width = 600,
  height = 600
)
}
\arguments{
\item{gpa}{Geomorph GPA object containing the consensus shape.}

\item{pca}{PCA object used to compute CAC.}

\item{cac_res}{Output of \code{CAC()} containing loadings and scores.}

\item{min_col}{Color for the minimal shape (default = "blue").}

\item{max_col}{Color for the maximal shape (default = "red").}

\item{mid_col}{Color for the mean shape (default = "black").}

\item{size_min, }{size_mid, size_max Point sizes for min/mean/max shapes.}

\item{size_mid}{Point size for landmarks in the consensus deformation plot.}

\item{size_max}{Point size for landmarks in the maximal deformation plot.}

\item{width, height}{Width and height (pixels) of the rgl 3D widget.}
}
\value{
An interactive rgl 3D widget displaying the shapes.
}
\description{
This function visualizes theoretical 3D shapes along a specified CAC axis,
showing minimal, maximal, and mean shapes in a 3D rgl viewer.
}
