% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pspline_smooth}
\alias{pspline_smooth}
\title{Smooths a 3D curve using P-splines}
\usage{
pspline_smooth(points, n_out = 100, k = 7)
}
\arguments{
\item{points}{A numeric matrix of size n × 3 representing the 3D coordinates of the curve.}

\item{n_out}{Integer. Number of output points along the smoothed curve. Default is 100.}

\item{k}{Integer. Number of basis functions for the spline (controls smoothness). Default is \code{7}.}
}
\value{
A numeric matrix of size n_out × 3 with smoothed X, Y, Z coordinates.
}
\description{
This function fits a cubic P-spline to each coordinate (X, Y, Z) of a 3D curve
and generates a smooth version of the curve with a specified number of points.
Use for smoothing noisy or/and resampling semi-landmark curves.
}
