% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reflect_point}
\alias{reflect_point}
\title{Reflects a point across a plane}
\usage{
reflect_point(X, origin, pc3)
}
\arguments{
\item{X}{Numeric vector of length 3. The point to reflect.}

\item{origin}{Numeric vector of length 3. A point on the plane.}

\item{pc3}{Numeric vector of length 3. Normal vector of the plane (should be normalized).}
}
\value{
Numeric vector of length 3 representing the reflected point.
}
\description{
This function computes the reflection of a 3D point/landmark across a plane.
The plane is defined by a point \code{origin} and a normal vector \code{pc3}.
}
\examples{
origin <- c(0, 0, 0)
normal <- c(0, 0, 1)   # plane XY
point <- c(1, 2, 3)
reflected <- reflect_point(point, origin, normal)
# reflected should be (1, 2, -3)
}
