% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{view_specimens}
\alias{view_specimens}
\title{Visualize 3D landmarks of ammonite specimens}
\usage{
view_specimens(landmarks_list, specimen_id = NULL, all = FALSE)
}
\arguments{
\item{landmarks_list}{A named list of specimens. Each specimen must be a list
of curves, where each curve is a matrix of 3D coordinates (X, Y, Z).}

\item{specimen_id}{A character string specifying the ID of a specimen to display.
Must match one of the names of \code{landmarks_list}. If \code{NULL} (default),
an interactive menu will be displayed unless \code{all = TRUE}.}

\item{all}{Logical, if \code{TRUE} all specimens in \code{landmarks_list} are displayed
sequentially. The user must press \code{Enter} to continue to the next specimen.}
}
\value{
This function is called for its side effect: opening an interactive
3D window via \pkg{rgl}. It does not return a value.
}
\description{
This function allows you to visualize one or several ammonite specimens
in 3D using \strong{rgl}. Curves are colored according to their type
(R = red, L = blue, U/V = green). You can either display a single specimen
(by ID), all specimens one after the other, or interactively choose from a menu.
}
\details{
Colors are assigned automatically based on the first letter of the curve name:
\itemize{
\item R = red
\item L = blue
\item U and V = green
\item Any other curve = black
}
}
\examples{
\dontrun{
# Display one specimen
view_specimens(data$landmarks, specimen_id = "182")

# Browse all specimens
view_specimens(data$landmarks, all = TRUE)

# Choose interactively
view_specimens(data$landmarks)
}

}
